Combine getHistory into getElement
