+++ handlers/element_version.js
@@ -2,7 +2,7 @@ var Readable = require('readable-stream').Readable
 var toOsm = require('obj2osm')
 
 module.exports = function (req, res, api, params, next) {
-  api.getElement(params.id, params.version, function (err, element) {
+  api.getElement(params.id, {version: params.version}, function (err, element) {
     if (err) return next(err)
     var r = new Readable({objectMode: true})
     r.on('error', next)
